services/api/pyproject.toml
[tool.poetry]
name = "minca-api"
version = "0.1.0"
description = "Minca AI Insurance API Service"
authors = ["Minca AI Team <team@mincaai.com>"]
packages = [{ include = "api", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.8.0"
sqlalchemy = "^2.0.30"
psycopg = {extras = ["binary"], version="^3.1.18"}
boto3 = "^1.34.0"
python-dotenv = "^1.0.1"
httpx = "^0.28.0"
# Workspace packages  
# Add missing dependencies for worker_transform import
pandas = "^2.0.0"
openpyxl = "^3.1.0"
PyYAML = "^6.0.0"
db = {path = "../../packages/db", develop = true}
storage = {path = "../../packages/storage", develop = true}
schemas = {path = "../../packages/schemas", develop = true}
mincaai-mq = {path = "../../packages/mq", develop = true}

[tool.poetry.group.dev.dependencies]
black = "^24.0"
ruff = "^0.5.0"
mypy = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"